+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### new
 * new task `getFilesDump` which will tar all files in the `filesFolder` into a tar and download it to the local computer.
+* `dockerKeyFile`, `dockerAuthorizedKeyFile` and `dockerKnownHostsFile` can now be loaded from http(s)
 
 ## 2.2.5