+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 2.x.x
 
 ### new
+* new task `sshCommand`: it will print out a ssh command to log into a remote instance. With some shell-magic you can streamline it into a single shell-command. This command will not have the downsides of the `ssh`-command e.g. the lagging shell.
 * new task `installFrom` which will do a install-task and a copyFrom afterwards. It will run the reset-task only after the copyFrom, so performance is better
 * new task `getFilesDump` which will tar all files in the `filesFolder` into a tar and download it to the local computer.
 * `dockerKeyFile`, `dockerAuthorizedKeyFile` and `dockerKnownHostsFile` can now be loaded from http(s)